backend/requirements/base.txt
# Backend requirements - base
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0
django-mptt>=0.16,<1.0
django-celery-beat>=2.5,<3.0

# Channels (WebSocket)
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Celery
celery>=5.3,<6.0
redis>=5.0,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Utils
Pillow>=10.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=21.0,<23.0
